+++ admin/delete_photo.php
@@ -4,24 +4,28 @@
 
 <?php
 
-    if(empty($_GET['id'])){
-        redirect("photos.php");
+    if(!User::find_by_id($_SESSION['user_id'])->is_admin() & $_SESSION['user_id'] != Photo::find_by_id($_GET['id'])->user_id){
+        redirect("photos_user.php");
+    }else{
+
+        if(empty($_GET['role'])){
+            $direct_to = "photos_user.php";
+        }else{
+            $direct_to = "photos.php";
+        }
+
+        if(empty($_GET['id'])){
+            redirect($direct_to);
+        }
+
+        $photo = Photo::find_by_id($_GET['id']);
+
+        if($photo){
+            $photo->delete_photo();
+        }
+
+        $session->message("The photo: {$photo->title} has been deleted");
+        redirect($direct_to);
     }
-
-    $photo = Photo::find_by_id($_GET['id']);
-
-    if($photo){
-        $photo->delete_photo();
-    }
-
-    $session->message("The photo: {$photo->title} has been deleted");
-    redirect("photos.php");
-    
-    // if($photo){
-    //     $photo->delete_photo();
-    //     redirect("../photos.php");
-    // }else{
-    //     redirect("../photos.php");
-    // }
 
 ?>
